pyproject.toml
[project]
name = "ocr_to_doc"
version = "0.1.0"
description = ""
authors = [
    {name = "t_ryohei"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "yomitoku (>=0.10.1,<0.11.0)",
    "pdf2image (>=1.17.0,<2.0.0)",
    "pillow-heif (>=1.1.1,<2.0.0)",
    "cairosvg (>=2.7.1,<3.0.0)",
    "python-docx (>=0.8.11,<1.0.0)",
    "markitdown[all] (>=0.1.3,<0.2.0)",
    "image2pdf (>=1.0.0,<2.0.0)",
    "pix2text[multilingual] (>=1.1.0,<2.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "pillow-heif (>=1.1.1,<2.0.0)",
    "pytesseract (>=0.3.13,<0.4.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
